Add PassportJS for user auth; Add workspace service

+++ src/resolvers/workspace/workspace.resolver.ts
@@ -1,57 +1,108 @@
-import { UseGuards } from '@nestjs/common';
 import {
   Args,
-  Context,
   Mutation,
   Parent,
   Query,
   ResolveField,
   Resolver,
 } from '@nestjs/graphql';
-import { GqlAuthGuard } from 'src/guards/auth';
+import { CurrentUser } from 'src/decorators';
 import { PrismaService } from 'src/prisma.service';
-import { Ctx } from 'src/types';
 import {
   CreateOneWorkspaceArgs,
   DeleteOneWorkspaceArgs,
   FindUniqueWorkspaceArgs,
   Task,
   UpdateOneWorkspaceArgs,
+  User,
   Workspace,
   WorkspaceMember,
+  WorkspaceMemberCreateWithoutWorkspaceInput,
 } from '../../../prisma/generated/types';
+import { WorkspaceService } from './workspace.service';
 
 @Resolver((of) => Workspace)
 export class WorkspaceResolver {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    private workspaceService: WorkspaceService,
+  ) {}
 
   @Query((returns) => [Workspace], { name: 'workspaces' })
-  findAll(@Context() context: Ctx) {
-    console.log('context', context.user);
-
-    return this.prisma.workspace.findMany();
+  findAll(@CurrentUser() user: User) {
+    return this.prisma.workspace.findMany({
+      where: {
+        members: {
+          some: {
+            userId: user.uuid,
+          },
+        },
+      },
+    });
   }
 
   @Query(() => Workspace, { name: 'workspace' })
-  findOne(@Args() findUniqueWorkspaceArgs: FindUniqueWorkspaceArgs) {
+  async findOne(
+    @CurrentUser() user: User,
+    @Args() findUniqueWorkspaceArgs: FindUniqueWorkspaceArgs,
+  ) {
+    await this.workspaceService.ensureUserIsWorkspaceMember(
+      user.email,
+      findUniqueWorkspaceArgs.where.uuid,
+    );
+
     return this.prisma.workspace.findUnique(findUniqueWorkspaceArgs);
   }
 
   @Mutation(() => Workspace)
-  createWorkspace(@Args() createWorkspaceInput: CreateOneWorkspaceArgs) {
-    return this.prisma.workspace.create(createWorkspaceInput);
+  createWorkspace(
+    @CurrentUser() user: User,
+    @Args() createWorkspaceInput: CreateOneWorkspaceArgs,
+  ) {
+    const { data } = createWorkspaceInput;
+    const OWNER: WorkspaceMemberCreateWithoutWorkspaceInput = {
+      role: 'OWNER',
+      user: {
+        connect: {
+          uuid: user.uuid,
+        },
+      },
+    };
+
+    return this.prisma.workspace.create({
+      data: {
+        ...data,
+        members: {
+          ...data.members,
+          create: [OWNER],
+        },
+      },
+    });
   }
 
   @Mutation(() => Workspace)
-  updateWorkspace(@Args() updateWorkspaceInput: UpdateOneWorkspaceArgs) {
+  async updateWorkspace(
+    @CurrentUser() user: User,
+    @Args() updateWorkspaceInput: UpdateOneWorkspaceArgs,
+  ) {
+    await this.workspaceService.ensureUserIsWorkspaceOwnerOrAdmin(
+      updateWorkspaceInput.where.uuid,
+      user.uuid,
+    );
     return this.prisma.workspace.update(updateWorkspaceInput);
   }
 
   @Mutation(() => Workspace)
-  removeWorkspace(
+  async removeWorkspace(
+    @CurrentUser() user: User,
     @Args()
     deleteOneWorkspaceArgs: DeleteOneWorkspaceArgs,
   ) {
+    await this.workspaceService.ensureUserIsWorkspaceOwner(
+      deleteOneWorkspaceArgs.where.uuid,
+      user.uuid,
+    );
+
     return this.prisma.workspace.delete(deleteOneWorkspaceArgs);
   }
